chore: update dependencies and improve settings/system integrations
- Update deps: notify 8.0, portable-pty 0.9, resvg/usvg 0.46, rusqlite 0.38, sysinfo 0.37
- Remove pulldown-cmark dependency (unused)
- Settings action now opens config.ts in user's configured editor with folder context
- Fix rusqlite 0.38 compatibility (i64 params for LIMIT/OFFSET, COUNT return type)
- Remove default notes hotkey - users must explicitly configure one
- Rewrite system preferences to use URL scheme for macOS Ventura+ compatibility

+++ src/app_launcher.rs
@@ -490,7 +490,7 @@ pub fn get_apps_db_stats() -> (usize, u64) {
         Err(_) => return (0, 0),
     };
 
-    let count: usize = conn
+    let count: i64 = conn
         .query_row("SELECT COUNT(*) FROM apps", [], |row| row.get(0))
         .unwrap_or(0);
 
@@ -502,7 +502,7 @@ pub fn get_apps_db_stats() -> (usize, u64) {
         )
         .unwrap_or(0);
 
-    (count, total_icon_size as u64)
+    (count as usize, total_icon_size as u64)
 }
 
 // ============================================================================
